Introduce Update Version in POM.
Experimental version only.
Works only for Applications, not for libs.
The basic problem with the libs is that there is not info.plist file that
holds the versions.

Unfortunatally it is hard to check which kind of project a project represents.
Therefore the corresponding menu entry is visible also for libs.